+++ CHANGELOG.md
@@ -1,6 +1,11 @@
-## 6.7.6
+## 6.7.6(Upcoming release)
+### Features
+- Add backup to mariadb
+## Refactor
+- Use MariaDB bundle product instead of sdk-go-dbaas-mariadb
 ### Chore
 - Remove nolint, add comments
+
 ## 6.7.5
 ## Refactor
 - Use SDK bundle for Object Storage Management